+++ packages/api/src/Domain/Request/Recovery/RecoveryKeyParamsRequestParams.ts
@@ -1,0 +1,6 @@
+export interface RecoveryKeyParamsRequestParams {
+  api_version: string
+  username: string
+  code_challenge: string
+  recovery_codes: string
+}