+++ packages/api/src/Domain/Response/Recovery/RecoveryKeyParamsResponseBody.ts
@@ -1,0 +1,5 @@
+import { KeyParamsData } from '@standardnotes/responses'
+
+export interface RecoveryKeyParamsResponseBody {
+  keyParams: KeyParamsData
+}