+++ packages/snjs/lib/Domain/UseCase/UseCaseContainerInterface.ts
@@ -1,0 +1,6 @@
+import { UseCaseInterface } from '@standardnotes/domain-core'
+
+export interface UseCaseContainerInterface {
+  get signInWithRecoveryCodes(): UseCaseInterface<void>
+  get getRecoveryCodes(): UseCaseInterface<string>
+}