+++ packages/desktop/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.104.80](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.136.7...@standardnotes/desktop@3.104.80) (2023-01-10)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.104.79](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.136.6...@standardnotes/desktop@3.104.79) (2023-01-10)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.104.78](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.136.5...@standardnotes/desktop@3.104.78) (2023-01-09)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.104.77](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.136.4...@standardnotes/desktop@3.104.77) (2023-01-09)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.104.76](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.136.3...@standardnotes/desktop@3.104.76) (2023-01-09)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
 ## [3.104.75](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.136.2...@standardnotes/desktop@3.104.75) (2023-01-07)
 
 **Note:** Version bump only for package @standardnotes/desktop