+++ packages/docs/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.8](https://github.com/standardnotes/app/compare/@standardnotes/docs@0.2.7...@standardnotes/docs@0.2.8) (2023-01-10)
+
+**Note:** Version bump only for package @standardnotes/docs
+
+## [0.2.7](https://github.com/standardnotes/app/compare/@standardnotes/docs@0.2.6...@standardnotes/docs@0.2.7) (2023-01-10)
+
+**Note:** Version bump only for package @standardnotes/docs
+
 ## [0.2.6](https://github.com/standardnotes/app/compare/@standardnotes/docs@0.2.5...@standardnotes/docs@0.2.6) (2023-01-03)
 
 **Note:** Version bump only for package @standardnotes/docs