+++ packages/releases/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.103](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.4.102...@standardnotes/releases@1.4.103) (2023-01-10)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.4.102](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.4.101...@standardnotes/releases@1.4.102) (2023-01-10)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.4.101](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.4.100...@standardnotes/releases@1.4.101) (2023-01-09)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.4.100](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.4.99...@standardnotes/releases@1.4.100) (2023-01-09)
+
+**Note:** Version bump only for package @standardnotes/releases
+
+## [1.4.99](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.4.98...@standardnotes/releases@1.4.99) (2023-01-09)
+
+**Note:** Version bump only for package @standardnotes/releases
+
 ## [1.4.98](https://github.com/standardnotes/app/compare/@standardnotes/releases@1.4.97...@standardnotes/releases@1.4.98) (2023-01-07)
 
 **Note:** Version bump only for package @standardnotes/releases